Fix warning about comparing values with different signedness (GCC9)

+++ encfs/encfs.cpp
@@ -729,7 +729,7 @@ int encfs_read(const char *path, char *buf, size_t size, off_t offset,
                struct fuse_file_info *file) {
   // Unfortunately we have to convert from ssize_t (pread) to int (fuse), so
   // let's check this will be OK
-  if (size > std::numeric_limits<int>::max()) {
+  if (size > (size_t)std::numeric_limits<int>::max()) {
     size = std::numeric_limits<int>::max();
   }
   return withFileNode("read", path, file,
@@ -757,7 +757,7 @@ int encfs_write(const char *path, const char *buf, size_t size, off_t offset,
                 struct fuse_file_info *file) {
   // Unfortunately we have to convert from ssize_t (pwrite) to int (fuse), so
   // let's check this will be OK
-  if (size > std::numeric_limits<int>::max()) {
+  if (size > (size_t)std::numeric_limits<int>::max()) {
     size = std::numeric_limits<int>::max();
   }
   EncFS_Context *ctx = context();
